Merge master to nmirror

+++ tests/runtest.cmd
@@ -71,15 +71,17 @@ for /f "delims=" %%a in ('dir /s /aD /b src\*') do (
     set __SourceFileName=%%~na
     set __RelativePath=!__SourceFolder:%CoreRT_TestRoot%=!
     if exist "!__SourceFolder!\project.json" (
-        %CoreRT_CliDir%\dotnet restore !__SourceFolder!
+        %CoreRT_CliDir%\dotnet restore --quiet --source "https://dotnet.myget.org/F/dotnet-core" "!__SourceFolder!"
 
         set __Mode=Jit
         call :CompileFile !__SourceFolder! !__SourceFileName! %__LogDir%\!__RelativePath!
         set /a __JitTotalTests=!__JitTotalTests!+1
 
-        set __Mode=Cpp
-        call :CompileFile !__SourceFolder! !__SourceFileName! %__LogDir%\!__RelativePath! --cpp
-        set /a __CppTotalTests=!__CppTotalTests!+1
+        if not exist "!__SourceFolder!\no_cpp" (
+            set __Mode=Cpp
+            call :CompileFile !__SourceFolder! !__SourceFileName! %__LogDir%\!__RelativePath! --cpp
+            set /a __CppTotalTests=!__CppTotalTests!+1
+        )
     )
 )
 set /a __CppFailedTests=%__CppTotalTests%-%__CppPassedTests%
@@ -132,17 +134,17 @@ goto :eof
     if not exist "!__CompileLogPath!" (mkdir !__CompileLogPath!)
     set __SourceFile=!__SourceFolder!\!__SourceFileName!
 
-    rmdir /s /q !__SourceFolder!\bin
-    rmdir /s /q !__SourceFolder!\obj
+    if exist "!__SourceFolder!\bin" rmdir /s /q !__SourceFolder!\bin
+    if exist "!__SourceFolder!\obj" rmdir /s /q !__SourceFolder!\obj
 
     setlocal
     set additionalCompilerFlags=
     if /i "%CoreRT_BuildType%" == "debug" (
-        if /i "%__Mode%" == "cpp" set additionalCompilerFlags=--cppcompilerflags /D_DEBUG
+        if /i "%__Mode%" == "cpp" set additionalCompilerFlags=--cppcompilerflags /MTd
     )
     REM TODO: Add AppDepSDK argument after CLI build picks up: PR dotnet/cli #336
     call "!VS140COMNTOOLS!\..\..\VC\vcvarsall.bat" %CoreRT_BuildArch%
-    "%CoreRT_CliDir%\dotnet" compile --native --ilcpath "%CoreRT_ToolchainDir%" !__ExtraCompileArgs! !__SourceFolder! -c %CoreRT_BuildType% %additionalCompilerFlags%
+    "%CoreRT_CliDir%\dotnet" build --native --runtime "win7-x64" --ilcpath "%CoreRT_ToolchainDir%" !__ExtraCompileArgs! !__SourceFolder! -c %CoreRT_BuildType% %additionalCompilerFlags%
     endlocal
 
     set __SavedErrorLevel=%ErrorLevel%
@@ -151,7 +153,7 @@ goto :eof
     if "%__SavedErrorLevel%"=="0" (
         echo.
         echo Running test !__SourceFileName!
-        call !__SourceFile!.cmd %CoreRT_BuildType%
+        call !__SourceFile!.cmd !__SourceFolder!\bin\%CoreRT_BuildType%\dnxcore50\win7-x64\native !__SourceFileName!.exe
         set __SavedErrorLevel=!ErrorLevel!
     )
